+++ Tests/RapidSpecs/ComparableSpecs/ClampedSpec.swift
@@ -31,5 +31,27 @@ final class ClampedSpec: QuickSpec {
                 expect("e".clamped(to: "a"..."d")).to(equal("d"))
             }
         }
+        
+        describe("the clamp(to:) method") {
+            it("does not change the value if it is in the given range") {
+                var seven = 7, point32 = 0.32
+                
+                seven.clamp(to: 6...12)
+                point32.clamp(to: 0.31...0.33)
+                
+                expect(seven).to(equal(7))
+                expect(point32).to(equal(0.32))
+            }
+            
+            it("clamps the value otherwise") {
+                var one = 1, e = "e"
+                
+                one.clamp(to: 5...7)
+                e.clamp(to: "a"..."d")
+                
+                expect(one).to(equal(5))
+                expect(e).to(equal("d"))
+            }
+        }
     }
 }